fix dup post routes
