+++ server/src/controllers/PostController.ts
@@ -1,54 +1,59 @@
 import { Request, Response } from 'express';
 import Post from '../models/Post';
-import Community from '../models/Community';
 import Marker from '../models/Marker';
+import Community from '../models/Community';
 
-/* Used ChatGPT to understand how to write the following functions */
-// The following functions are used to get all posts, delete a post, and create a post.
+export const getPost = async (req: Request, res: Response) => {
+  const post = await Post.findById(req.params.postId).populate('replies');
 
-// getAllPosts from a specific community function
-export const getCommunityPosts = async (req: Request, res: Response) => {
-  const { communityId } = req.params;
-
-  try {
-    const posts = await Post.find(
-      { community: communityId },
-      'user likeCount type marker content'
-    ).populate([
-      { path: 'user', select: 'userName firstName lastName' },
-      { path: 'marker' },
-    ]);
-
-    res.status(200).json(posts);
-  } catch (error) {
-    console.error('Get Posts Error:', error);
-    res.statusMessage = error as string;
-    return res.sendStatus(400);
+  if (!post) {
+    return res.status(404).json({ message: 'Post not found' });
   }
+
+  res.json(post);
 };
 
-// deletePost function
-export const deletePost = async (req: Request, res: Response) => {
-  const { id } = req.params;
+export const getReplies = async (req: Request, res: Response) => {
+  const post = await Post.findById(req.params.postId).populate('replies');
 
-  try {
-    if (!id) {
-      res.statusMessage = 'Bad request, missing post ID.';
-      return res.sendStatus(400);
-    }
-
-    await Post.findByIdAndDelete(id);
-
-    res.sendStatus(204); // Returns no content, 204 status meaning success.
-  } catch (error) {
-    console.error('Delete Post Error:', error);
-    res.statusMessage = error as string;
-    return res.sendStatus(400);
+  if (!post) {
+    return res.status(404).json({ message: 'Post not found' });
   }
+
+  res.json(post.replies);
+};
+
+export const addReply = async (req: Request, res: Response) => {
+  const post = await Post.findById(req.params.postId);
+
+  if (!post) {
+    return res.status(404).json({ message: 'Post not found' });
+  }
+
+  post.replies.push(req.body.replyId);
+  await post.save();
+  res.json(post);
+};
+
+export const deleteReply = async (req: Request, res: Response) => {
+  const post = await Post.findById(req.params.postId);
+
+  if (!post) {
+    return res.status(404).json({ message: 'Post not found' });
+  }
+
+  const replyIndex = post.replies.findIndex(
+    (reply) => reply.toString() === req.params.replyId
+  );
+  if (replyIndex !== -1) {
+    post.replies.splice(replyIndex, 1);
+    await post.save();
+  }
+  res.json({ message: 'Reply deleted' });
 };
 
 // createPost function
-export const savePost = async (req: Request, res: Response) => {
+export const createPost = async (req: Request, res: Response) => {
   const { community, user, type, content, marker } = req.body;
 
   try {
@@ -91,6 +96,26 @@ export const savePost = async (req: Request, res: Response) => {
     res.status(201).json({ data: post });
   } catch (error) {
     console.error('Create Post Error:', error);
+    res.statusMessage = error as string;
+    return res.sendStatus(400);
+  }
+};
+
+// deletePost function
+export const deletePost = async (req: Request, res: Response) => {
+  const { id } = req.params;
+
+  try {
+    if (!id) {
+      res.statusMessage = 'Bad request, missing post ID.';
+      return res.sendStatus(400);
+    }
+
+    await Post.findByIdAndDelete(id);
+
+    res.sendStatus(204); // Returns no content, 204 status meaning success.
+  } catch (error) {
+    console.error('Delete Post Error:', error);
     res.statusMessage = error as string;
     return res.sendStatus(400);
   }
