+++ server/src/routes/Posts.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import { validateToken } from '../middlewares';
-import * as postController from '../controllers/PostControllers';
+import * as postController from '../controllers/PostController';
 
 const router = express.Router();
 
@@ -11,9 +11,14 @@ router.get('/:postId/replies', validateToken, postController.getReplies);
 
 router.post('/:postId/replies', validateToken, postController.addReply);
 
-router.delete('/:postId/replies/:replyId', validateToken, postController.deleteReply);
+router.delete(
+  '/:postId/replies/:replyId',
+  validateToken,
+  postController.deleteReply
+);
 
 router.post('/', validateToken, postController.createPost);
 
+router.delete('/:id', validateToken, postController.deletePost);
 
 export default router;
